refactor: separate runner execution layout
